frontend/SG_Predictor/requirement.txt
pandas
numpy
yfinance
psycopg2-binary
scikit-learn